refactor(allocator): extract routes/provisioning.py from main.py
Move /api/launch, /destroy, /api/operations, and
/api/operations/<int:operation_id> (plus their shared _wants_json helper)
out of main.py into a new provisioning blueprint. The launch/destroy
closures run on OperationsWorker's background thread, so they read the
provider from current_app.config in the request thread only, and never
touch current_app themselves.
